+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-10 11:45:55
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-05-10 12:09:19
+// Commit: feat(ui): add async ui stream provider
 
 class uiComponentState {
-  final String commitRef = '992';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-05-10T11:45:55+05:00');
+  final String commitRef = '994';
+  final String lastMessage = 'feat(ui): add async ui stream provider';
+  final DateTime timestamp = DateTime.parse('2026-05-10T12:09:19+05:00');
 }